requirements.txt
load_dotenv
pyTelegramBotAPI
g4f[all]